Relocate admin components into separate folder Fix typo in nodemon config
