+++ app/frontend/src/components/admin/Clients.vue
