+++ app/frontend/src/components/admin/HealthDialog.vue
