+++ app/frontend/src/components/admin/JWTDialog.vue
