+++ app/nodemon.json
@@ -1,8 +1,8 @@
 {
   "ignore": [
     "frontend/src",
-    "frontend/test",
+    "frontend/tests",
     "node_modules/**/node_modules",
-    "test"
+    "tests"
   ]
 }
